[docs] 동기화 문제해결..
AOP와 충돌로 인한 동기화문제 해결 (proxyTargetClass = true)
