+++ src/main/java/com/team/final8teamproject/Final8TeamProjectApplication.java
@@ -4,6 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
+import org.springframework.core.Ordered;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -12,9 +13,9 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableJpaAuditing
 @EnableScheduling
 @EnableAspectJAutoProxy
-@EnableAsync
-@EnableCaching
-//(order = Ordered.HIGHEST_PRECEDENCE)
+@EnableAsync(proxyTargetClass=true)
+@EnableCaching(order = Ordered.HIGHEST_PRECEDENCE)
+//
 public class Final8TeamProjectApplication {
     public static void main(String[] args) {
         SpringApplication.run(Final8TeamProjectApplication.class, args);
