+++ src/main/java/com/team/final8teamproject/security/service/EmailService.java
@@ -1,6 +1,7 @@
 package com.team.final8teamproject.security.service;
 
 import com.team.final8teamproject.share.exception.CustomException;
+import com.team.final8teamproject.user.dto.FindPasswordRequestDto;
 
 import java.time.LocalDateTime;
 
@@ -10,4 +11,6 @@ public interface EmailService {
     boolean verifyAuthCode(String to, String authCode) throws CustomException;
 
     LocalDateTime getAuthCodeCreatedAt(String to);
+
+    void sendEmail(FindPasswordRequestDto vo, String password);
 }
