+++ src/main/java/com/team/final8teamproject/security/service/EmailServiceImpl.java
@@ -2,12 +2,15 @@ package com.team.final8teamproject.security.service;
 
 import com.team.final8teamproject.share.exception.CustomException;
 import com.team.final8teamproject.share.exception.ExceptionStatus;
+import com.team.final8teamproject.user.dto.FindPasswordRequestDto;
 import jakarta.mail.Message;
+import jakarta.mail.MessagingException;
 import jakarta.mail.internet.InternetAddress;
 import jakarta.mail.internet.MimeMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -18,7 +21,6 @@ import java.util.Random;
 @Service
 @RequiredArgsConstructor
 public class EmailServiceImpl implements EmailService {
-
     private final JavaMailSender emailSender;
     private Map<String, String> authCodes = new HashMap<>();
     private Map<String, LocalDateTime> authCodeExpirationTimes = new HashMap<>();
@@ -84,6 +86,7 @@ public class EmailServiceImpl implements EmailService {
     }
 
     @Override
+    @Async
     public void sendSimpleMessage(String to) throws Exception {
         String ePw = createKey();
         authCodes.put(to, ePw); // 이메일과 인증 코드를 Map에 저장
@@ -115,7 +118,32 @@ public class EmailServiceImpl implements EmailService {
     }
 
     public LocalDateTime getAuthCodeCreatedAt(String to) {
+
         return authCodeCreatedAt.get(to);
     }
 
+    // 유저 비밀번호 변경
+    //5-1.이메일 발송
+    @Async
+    public void sendEmail(FindPasswordRequestDto vo, String password) {
+        // Mail Server 설정
+        MimeMessage mail = emailSender.createMimeMessage();
+        // 받는 사람 E-Mail 주소
+        String userMail = vo.getEmail();
+        String htmlStr = "<h2>안녕하세요 '" + vo.getUsername() + "' 님</h2><br><br>"
+                + "<p>비밀번호 찾기를 신청해주셔서 임시 비밀번호를 발급해드렸습니다.</p>"
+                + "<p>임시로 발급 드린 비밀번호는 <h2 style='color : blue'>'" + password
+                + "'</h2>이며 로그인 후 마이페이지에서 비밀번호를 변경해주시면 됩니다.</p><br>"
+                + "<h3><a href='http://localhost:5500/index.html'>MS :p 홈페이지 접속 ^0^</a></h3><br><br>"
+                + "(혹시 잘못 전달된 메일이라면 이 이메일을 무시하셔도 됩니다)";
+
+        try {
+            mail.setSubject("[MS :p] 임시 비밀번호가 발급되었습니다", "utf-8");
+            mail.setText(htmlStr, "utf-8", "html");
+            mail.addRecipient(MimeMessage.RecipientType.TO, new InternetAddress(userMail));
+            emailSender.send(mail);
+        } catch (MessagingException e) {
+            e.printStackTrace();
+        }
+    }
 }
