+++ src/main/java/com/team/final8teamproject/user/service/UserService.java
@@ -4,6 +4,7 @@ import com.team.final8teamproject.base.entity.BaseEntity;
 import com.team.final8teamproject.base.repository.BaseRepository;
 import com.team.final8teamproject.redis.cache.CacheNames;
 import com.team.final8teamproject.redis.RedisUtil;
+import com.team.final8teamproject.security.service.EmailService;
 import com.team.final8teamproject.share.exception.CustomException;
 import com.team.final8teamproject.share.exception.ExceptionStatus;
 import com.team.final8teamproject.user.dto.*;
@@ -40,13 +41,13 @@ import org.slf4j.Logger;
 @RequiredArgsConstructor
 public class UserService {
 
-    private final JavaMailSender mailSender;
     private final UserRepository userRepository;
     private final BaseRepository baseRepository;
     private final RedisUtil redisUtil;
     private final JwtUtil jwtUtil;
     private final PasswordEncoder passwordEncoder;
 
+    private final EmailService emailService;
     private final Logger loggers = LoggerFactory.getLogger(UserService.class);
     private final RedisTemplate<String, String> redisTemplate;
     //1. 회원가입
@@ -143,12 +144,12 @@ public class UserService {
                 () -> new IllegalArgumentException("이메일을 다시 입력해주시기 바랍니다.")
         );
         // 가입된 아이디가 없으면
-        if (user.isUsername(vo.getUsername())) {
+        if (!user.isUsername(vo.getUsername())) {
             throw new IllegalArgumentException("등록되지 않은 사용자입니다.");
         }
 
         // 가입된 이메일이 아니면
-        else if (user.isEmail(vo.getEmail())) {
+        else if (!user.isEmail(vo.getEmail())) {
             throw new IllegalArgumentException("등록되지 않은 사용자입니다.");
         } else {
 
@@ -159,35 +160,12 @@ public class UserService {
             }
             user.changePassword(passwordEncoder.encode(pw.toString()));
             // 비밀번호 변경 메일 발송
-            sendEmail(vo, pw.toString());
+            emailService.sendEmail(vo, pw.toString());
 
         }
         return new FindByResponseDto("임시 패스워드 발송 성공");
     }
 
-    //5-1.이메일 발송
-    @Async
-    public void sendEmail(FindPasswordRequestDto vo, String password) {
-        // Mail Server 설정
-        MimeMessage mail = mailSender.createMimeMessage();
-        // 받는 사람 E-Mail 주소
-        String userMail = vo.getEmail();
-        String htmlStr = "<h2>안녕하세요 '" + vo.getUsername() + "' 님</h2><br><br>"
-                + "<p>비밀번호 찾기를 신청해주셔서 임시 비밀번호를 발급해드렸습니다.</p>"
-                + "<p>임시로 발급 드린 비밀번호는 <h2 style='color : blue'>'" + password
-                + "'</h2>이며 로그인 후 마이페이지에서 비밀번호를 변경해주시면 됩니다.</p><br>"
-                + "<h3><a href='http://localhost:5500/index.html'>MS :p 홈페이지 접속 ^0^</a></h3><br><br>"
-                + "(혹시 잘못 전달된 메일이라면 이 이메일을 무시하셔도 됩니다)";
-
-        try {
-            mail.setSubject("[MS :p] 임시 비밀번호가 발급되었습니다", "utf-8");
-            mail.setText(htmlStr, "utf-8", "html");
-            mail.addRecipient(MimeMessage.RecipientType.TO, new InternetAddress(userMail));
-            mailSender.send(mail);
-        } catch (MessagingException e) {
-            e.printStackTrace();
-        }
-    }
     //6. 엑세스 토큰 재발급
 
     /**
